requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
python-decouple>=3.8
requests>=2.31.0
groq>=0.4.0
gunicorn>=21.2.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
